backend/requirements.txt
Flask
Flask-CORS
python-dotenv
requests
supabase
groq
email-validator
pytest
pytest-flask
smtplib
PyPDF2
python-docx
